requirements.txt
Django>=4.2,<6.0
azure-storage-blob>=12.19.0
python-dotenv>=1.0.1
mutagen>=1.47.0
psycopg2-binary>=2.9.9
gunicorn>=21.2.0
whitenoise>=6.5.0
whitenoise>=6.6.0
